pyproject.toml
[tool.poetry]
name = "bot-telegram"
version = "0.1.0"
description = ""
authors = ["Rodrigo <rodriguinhobettiojr1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytelegrambotapi = "^4.20.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
bot_telegram = "bot_telegram.main:main"

[tool.poetry.env]
CHAVE_API = "S7391313820:AAHdpnyJrTet65TwC7vG7u4C-L2Zg2p5TOo"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
